MAILBOX-64 Moving Quota related interfaces to a new package
