+++ api/src/main/java/org/apache/james/mailbox/quota/QuotaManager.java
@@ -16,8 +16,9 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.mailbox;
+package org.apache.james.mailbox.quota;
 
+import org.apache.james.mailbox.MailboxSession;
 import org.apache.james.mailbox.exception.MailboxException;
 import org.apache.james.mailbox.model.Quota;
